refactor: centralize type definitions
