+++ iwfpapp/lib/screens/add_card_from_template/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
 import 'package:iwfpapp/services/config/typedefs/home_tab_id.dart';
-import 'package:iwfpapp/services/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/submission_screen_status.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'package:iwfpapp/services/status.dart';
 import 'package:iwfpapp/services/utilities/card_template.dart';
 import 'package:iwfpapp/widgets/credit_cards/template.dart';
