+++ iwfpapp/lib/screens/add_promo/main.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:iwfpapp/services/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/shop_category.dart';
+import 'package:iwfpapp/services/config/typedefs/submission_screen_status.dart';
 import 'package:iwfpapp/services/config/typedefs/validation_response.dart';
-import 'package:iwfpapp/services/credit_card.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'package:iwfpapp/services/shop_category.dart';
-import 'package:iwfpapp/services/status.dart';
 import 'package:iwfpapp/services/utilities/validators/promo_info_validator.dart';
 
 class AddPromoScreen extends StatefulWidget {
