+++ iwfpapp/lib/screens/edit_card/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
 import 'package:iwfpapp/services/config/typedefs/home_tab_id.dart';
-import 'package:iwfpapp/services/credit_card.dart';
-import 'package:iwfpapp/services/status.dart';
+import 'package:iwfpapp/services/config/typedefs/submission_screen_status.dart';
 import 'package:iwfpapp/widgets/promos/entry_view.dart';
 
 class EditCardScreen extends StatefulWidget {
