+++ iwfpapp/lib/screens/remove_promo/main.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
 import 'package:iwfpapp/services/config/typedefs/remove_promo.dart';
-import 'package:iwfpapp/services/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/submission_screen_status.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'package:iwfpapp/services/status.dart';
 
 class RemovePromoScreen extends StatefulWidget {
   final DataStore dataStore;
