+++ iwfpapp/lib/screens/sign_up/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:iwfpapp/services/auth.dart';
 import 'package:iwfpapp/services/config/typedefs/auth_status.dart';
-import 'package:iwfpapp/services/status.dart';
+import 'package:iwfpapp/services/config/typedefs/submission_screen_status.dart';
 
 class SignUpScreen extends StatefulWidget {
   final IwfpappAuth auth;
