+++ iwfpapp/lib/screens/suggestion/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:iwfpapp/services/config/typedefs/home_tab_id.dart';
+import 'package:iwfpapp/services/config/typedefs/shop_category.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'package:iwfpapp/services/shop_category.dart';
 import 'package:iwfpapp/screens/suggestion/entries.dart';
 
 class SuggestionScreen extends StatelessWidget {
