+++ iwfpapp/lib/services/config/typedefs/credit_card.dart
@@ -1,4 +1,4 @@
-import 'package:iwfpapp/services/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
 
 class CreditCard {
   final String name;
