+++ iwfpapp/lib/services/config/typedefs/remove_promo.dart
@@ -1,5 +1,5 @@
-import 'package:iwfpapp/services/cashback_promo.dart';
-import 'package:iwfpapp/services/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
 
 class RemovePromoMeta {
   final CreditCard card;
