+++ iwfpapp/lib/services/config/typedefs/shop_category.dart
