+++ iwfpapp/lib/services/ranker.dart
@@ -1,9 +1,9 @@
 import 'dart:math';
 
-import 'package:iwfpapp/services/cashback_promo.dart';
-import 'package:iwfpapp/services/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/shop_category.dart';
 import 'package:iwfpapp/services/fetcher.dart';
-import 'package:iwfpapp/services/shop_category.dart';
 
 bool isInValidTimeRange(CashbackPromo promo, {DateTime useCurrentTime}) {
   bool valid = false;
