+++ iwfpapp/lib/services/utilities/category_counter.dart
@@ -1,6 +1,6 @@
-import 'package:iwfpapp/services/cashback_promo.dart';
-import 'package:iwfpapp/services/credit_card.dart';
-import 'package:iwfpapp/services/shop_category.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/shop_category.dart';
 
 List<ShopCategory> getUniqueShoppingCategories(List<CreditCard> cards) {
   Map<String, bool> promoIdDict = {};
