+++ iwfpapp/lib/services/utilities/validators/card_expiration_validator.dart
@@ -1,4 +1,4 @@
-import 'package:iwfpapp/services/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
 
 bool isInValidTimeRange(CashbackPromo promo, {DateTime useCurrentTime}) {
   bool valid = false;
