+++ iwfpapp/test/integration/mock_services/mock_data_store.dart
@@ -1,7 +1,7 @@
-import 'package:iwfpapp/services/cashback_promo.dart';
-import 'package:iwfpapp/services/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/cashback_promo.dart';
+import 'package:iwfpapp/services/config/typedefs/credit_card.dart';
+import 'package:iwfpapp/services/config/typedefs/shop_category.dart';
 import 'package:iwfpapp/services/data_store.dart';
-import 'package:iwfpapp/services/shop_category.dart';
 import 'package:iwfpapp/services/utilities/category_counter.dart';
 import 'package:mockito/mockito.dart';
